Rename the improper "vdiv" setting to "vscale".
The vertical scale setting is improperly named "vdiv" and as
such it creates a great deal of confusion especially in new
users and developers because it is easily mistaken with
"num_vdiv" which is the number of vertical divisions !

Appropriate measures are also introduced so that the vertical
scale unit is correctly printed as "V/div" instead of "V" (a
very small companion patch has been created for sigrok-cli).

It is up to individual driver maintainers to fix the local
variable names and comments (such source-cosmetic fixes are
provided here only for the "rohde-schwarz-hameg" driver).