Remove option bounds checking in the Rohde&Schwarz/Hameg driver.
With the recent introduction of the Event Status Register (ESR)
checking functionality, a very specific SCPI error (Execution
Error) is produced every time the user attempts to configure the
device with out of bounds values for some option.

Therefore, all currently hard-coded, sometimes model-dependent
bound checks are removed because the device already checks the
values and eventually returns an invalid ESR status.

Refresh the model state if the ESR contains SCPI command failures
to avoid an inconsistent state, for example, after trying to
set an out of bound option (device sets it exactly at the bound).