git-version.h: adjust header file creation for out-of-source builds
The git-version.h source file is generated output. Create the file
in $builddir so that $srcdir is not touched for out-of-source build
configurations. In-source builds won't change their behaviour since
$builddir equals $srcdir in that case.

Change to the $srcdir directory in another git(1) invocation, to handle
the case where the build directory is not a child of the source tree.