Makefile.am: make CLEAN_EXTRA available earlier
Move the CLEAN_EXTRA "initialization" to an earlier location in the
Makefile.am file. It's desirable to have all steps just append to the
variable, so that no incarnation needs to encode whether it's the first
reference to the variable. Unfortunately the .am syntax is rather picky.

This prepares a future commit where version control related steps want
to register extra 'make clean' targets, too.